Initial completion of Employee Management application.

+++ EmpManApp/src/main/java/com/ems/empmanagamentapp/service/EmployeeService.java
@@ -1,34 +1,61 @@
 package com.ems.empmanagamentapp.service;
+
 import com.ems.empmanagamentapp.model.Employee;
 import com.ems.empmanagamentapp.repository.EmployeeRepository;
+import com.ems.empmanagamentapp.service.interfaces.EmployeeServiceInterface;
 import org.springframework.stereotype.Service;
 import java.util.List;
+// import org.springframework.beans.factory.annotation.Autowired;
 
 @Service
-public class EmployeeService {
+public class EmployeeService implements EmployeeServiceInterface {
+
+    // @Autowired
     private final EmployeeRepository employeeRepository;
 
     public EmployeeService(EmployeeRepository employeeRepository) {
         this.employeeRepository = employeeRepository;
     }
 
+    @Override
+    public Employee createEmployee(Employee employee) {
+        return employeeRepository.save(employee);
+    }
+
+    @Override
     public List<Employee> getAllEmployees() {
         return employeeRepository.findAll();
     }
 
-    public Employee getEmployeeById(int id) {
-        return employeeRepository.findById(id);
+    @Override
+    public Employee getEmployeeById(Integer id) {
+        return employeeRepository.findById(id).orElseThrow(() -> new RuntimeException("Employee not found with ID: " + id));
     }
 
-    public void addEmployee(Employee employee) {
-        employeeRepository.save(employee);
+    @Override
+    public Employee updateEmployee(Integer id, Employee employee) {
+        if (employeeRepository.existsById(id)) {
+            employee.setId(id);
+            return employeeRepository.save(employee);
+        }
+        throw new RuntimeException("Employee not found with ID: " + id);
     }
 
-    public void updateEmployee(Employee employee) {
-        employeeRepository.update(employee);
+    @Override
+    public void deleteEmployee(Integer id) {
+        if (employeeRepository.existsById(id)) {
+            employeeRepository.deleteById(id);
+            return;
+        }
+        throw new RuntimeException("Employee not found with ID: " + id);
     }
 
-    public void deleteEmployee(int id) {
-        employeeRepository.deleteById(id);
+    @Override
+    public List<Employee> getEmployeesByDepartmentId(Integer departmentId) {
+        if (!employeeRepository.existsById(departmentId)) {
+            throw new RuntimeException("Department not found with ID: " + departmentId);
+        }
+        return employeeRepository.findByDepartmentId(departmentId);
     }
+
 }
